OCM-7254 | feat: Move edit machinepool non-cmd funcs to pkg and test
