+++ pkg/ocm/output/nodepools_test.go
@@ -1,8 +1,6 @@
 package output
 
 import (
-	"fmt"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
@@ -45,21 +43,4 @@ var _ = Describe("PrintNodePoolReplicasInline", func() {
 		Expect(output).To(Equal("2"))
 	})
 
-})
-
-var _ = Describe("Validate management upgrade print output", func() {
-	mgmtUpgrade, _ := cmv1.NewNodePoolManagementUpgrade().MaxSurge("10").MaxUnavailable("5").Type("Replace").Build()
-	DescribeTable("Validate management upgrade print output",
-		func(upgrade *cmv1.NodePoolManagementUpgrade, expectedOutput string) {
-			output := PrintNodePoolManagementUpgrade(upgrade)
-			Expect(output).To(Equal(expectedOutput))
-		},
-		Entry("Should return empty string", nil,
-			"",
-		),
-		Entry("Should return string with type, maxSurge and maxUnavailable",
-			mgmtUpgrade,
-			fmt.Sprintf("\n - Type:%38s\n - Max surge:%28s\n - Max unavailable:%21s", "Replace", "10", "5"),
-		),
-	)
 })
